refactor(tanstack-start): collapse _payload shims to package factories
